TransitionMgr is now a member of class StateMgr
This is an attempt to simplify code and reduce coupling between
all different classes which used to rely on both statemgr and
transitionmgr. As an added benefit, it should be easier now to
deduce when a transition must be created (logic will now only be
embedded in StateMgr class and not the callers of this class).