Simplify Graphviz plot by using specific font instead of pictures
After having discovered that a font was developed to represent all
signs discovered in the Indus valley so far, let's use this font
instead of the jpeg pictures previously used. This makes for a more
scalable graph (SVG font). The font itself can be downloaded for
free from the following url:

https://www.mohenjodaroonline.net/index.php/indus-script/corpus-by-asko-parpola/42-indus-script